LTS for 1D implementation and other fixes
