+++ test/helpers/CMakeLists.txt
@@ -14,5 +14,5 @@ add_library(testHelpers
 	"GlobalFunctions.cpp"
 )
 
-target_link_libraries(testHelpers Eigen3::Eigen)
+target_link_libraries(testHelpers maxwell Eigen3::Eigen)
