+++ test/mfem/CMakeLists.txt
@@ -18,6 +18,7 @@ add_executable(mfem_tests
 
 target_link_libraries(mfem_tests 
 	mfem 
+	testHelpers
 	GTest::gtest GTest::gtest_main
 )
